packages/sdk-python/pyproject.toml
[tool.poetry]
name = "parallax"
version = "0.1.0"
description = "Python SDK for Parallax AI orchestration platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "parallax", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.68.0"
grpcio-tools = "^1.68.0"
typing-extensions = "^4.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
ruff = "^0.9.1"

[tool.ruff]
target-version = "py39"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"